+++ docs/changes.rst
@@ -11,6 +11,8 @@ To be released.
   when you can't use :keyword:`with` keyword e.g. setup/teardown hooks.
 - :attr:`Image.object_type <sqlalchemy_imageattch.entity.Image.object_type>`
   property now has the default value when the primary key is an integer.
+- Columns of :class:`~sqlalchemy_imageattach.entity.Image` class become
+  able to be used as SQL expressions.
 
 
 Version 0.8.0.dev-20130531